Renamed release-8.02 (/branches/release-8.02 ) to release-08.02.

+++ src/plugins/contrib/wxSmith/wxwidgets/defitems/wxschoice.h
@@ -39,7 +39,7 @@ class wxsChoice: public wxsWidget
         virtual void OnEnumWidgetProperties(long Flags);
 
         wxArrayString ArrayChoices;
-        long DefaultSelection;
+        int DefaultSelection;
 };
 
 #endif
