+++ src/plugins/contrib/wxSmith/wxwidgets/defitems/wxscombobox.h
@@ -1,5 +1,4 @@
-/** \file wxscombobox.h
-*
+/*
 * This file is part of wxSmith plugin for Code::Blocks Studio
 * Copyright (C) 2006-2007  Bartlomiej Swiecki
 *
@@ -41,7 +40,7 @@ class wxsComboBox: public wxsWidget
 
 
         wxArrayString ArrayChoices;
-        long DefaultSelection;
+        int DefaultSelection;
 };
 
 #endif
