+++ src/plugins/contrib/wxSmith/wxwidgets/defitems/wxslistbox.h
@@ -40,7 +40,7 @@ class wxsListBox: public wxsWidget
 
 
         wxArrayString ArrayChoices;
-        long DefaultSelection;
+        int DefaultSelection;
 };
 
 #endif
